+++ macros/framework.rs
@@ -151,12 +151,12 @@ pub trait ExpansionOutput: SubstParseContext {
         A: FnOnce(&mut TokenAccumulator),
         B: FnOnce(&mut TokenAccumulator);
 
-    /// Append a [`syn::Lit`](enum@syn::Lit)
+    /// Append a [`syn::LitStr`](struct@syn::LitStr)
     ///
-    /// This is its own method because `syn::Lit` is not `Display`,
+    /// This is its own method because `syn::LitStr` is not `Display`,
     /// and we don't want to unconditionally turn it into a string
     /// before retokenising it.
-    fn append_syn_lit(&mut self, v: &syn::Lit);
+    fn append_syn_litstr(&mut self, v: &syn::LitStr);
 
     /// Append a [`syn::Type`]
     ///
@@ -164,15 +164,6 @@ pub trait ExpansionOutput: SubstParseContext {
     /// in the terminology of the template reference:
     /// If a paste contains more than one, it is an error.
     fn append_syn_type(&mut self, te_span: Span, v: &syn::Type);
-
-    /// Append a meta item value (without `as` clause in the template)
-    ///
-    /// Can fail, if the actual concrete value is not right
-    fn append_meta_value(
-        &mut self,
-        tspan: Span,
-        lit: &syn::Lit,
-    ) -> syn::Result<()>;
 
     /// Append using a function which generates tokens
     ///
@@ -408,20 +399,11 @@ impl ExpansionOutput for TokenAccumulator {
         self.append(ident);
         post(self);
     }
-    fn append_syn_lit(&mut self, lit: &syn::Lit) {
+    fn append_syn_litstr(&mut self, lit: &syn::LitStr) {
         self.append(lit);
     }
     fn append_syn_type(&mut self, _te_span: Span, ty: &syn::Type) {
         self.append(ty);
-    }
-    fn append_meta_value(
-        &mut self,
-        tspan: Span,
-        lit: &syn::Lit,
-    ) -> syn::Result<()> {
-        let tokens: TokenStream = metavalue_lit_as(lit, tspan, &"tokens")?;
-        self.append(tokens);
-        Ok(())
     }
     fn append_tokens_with(
         &mut self,